FER/requirements.txt
tensorflow>=2.10.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
Pillow>=8.3.0
scipy>=1.7.0 